Improve syscalls for coreutils including Output of entire files, Formatting file contents and Output of parts of files

+++ kernel/fs/vfs.h
@@ -68,7 +68,7 @@ typedef struct {
 } dirent_t;
 
 /* VFS data structure definitions */
-typedef int vfs_handle_t;
+typedef int64_t vfs_handle_t;
 
 /* Forward declaration */
 typedef struct vfs_inode_t vfs_inode_t;
